:truck: rename: 카카오 response 패키지 분리 #9
